Added index to users on name to speed up task creation Ran some basic benchamrks (benchmark/rails) on rails app

+++ benchmark/rails/benchmark.pl
@@ -8,7 +8,7 @@ $gHttperf = 'httperf';
 sub run_test {
     my($num_tasks, $rate, $timeout) = @_;
     $cmd = "$gHttperf --server=$gServer --method=POST --uri=/tasks/Learner+1/1/1 --num-conns=$num_tasks --rate=$rate --timeout=$timeout 2>/dev/null";
-    warn "Running '$cmd'\n";
+    warn "Running '$cmd'\n" if $gDebug;
     $output = `$cmd`;
     @results = &parse_results($output);
     return @results;
@@ -34,6 +34,10 @@ sub parse_results {
     } else {
         push(@errors, "Can't parse timeout");
     }
+    if ($total != $req || $total != $rep+$timeout) {
+        my $missing = $req - ($rep+$timeout);
+        push(@errors, "$total requests, but $req outgoing/$rep replies/$timeout T/O: $missing missing?");
+    }
     if (@errors) {
         my($fh,$name) = mkstemp("bmXXXXX");
         warn "Errors written to $name";
@@ -42,26 +46,23 @@ sub parse_results {
         close $fh;
     }
     return ($min/1000.0,$max/1000.0,$median/1000.0,$stdev/1000.0,
-            int(100.0*$timeout/$total),$dur);
+            100.0*$timeout/$total,$dur);
 }
 
 sub main {
-    $min_rate = 10;
-    $max_rate = 16;
-    $rate_step = 1;
-    $runs = 4;
+    die "Usage: $0 min_rate max_rate step num_runs_per_rate\n" unless $#ARGV == 3;
+    my($min_rate, $max_rate, $rate_step, $runs) = @ARGV;
     $tasks_to_create = 1000;
     $clear = "NEWRELIC_AGENT_ENABLED=false heroku run --app $gAppname rails runner -e production Task.delete_all >/dev/null 2>&1";
     $timeout = 10;
-    open(OUTPUT, "> output.txt") or die;
-    $out = OUTPUT;
+    $out = STDOUT;
     printf $out "| run | rate| min | max | med |stdv |fail%| dur |\n";
     for ($rate = $min_rate; $rate <= $max_rate; $rate += $rate_step) {
         for ($run = 1; $run <= $runs; $run++) {
             @result = &run_test($tasks_to_create, $rate, $timeout);
-            printf $out ("| %3d | %3d | %2.1f | %2.1f | %2.1f | %2.1f | %2d  |%4.0f |\n", $run, $rate, @result);
+            printf $out ("| %3d | %3d | %2.1f | %2.1f | %2.1f | %2.1f | %2.0f  |%4.0f |\n", $run, $rate, @result);
         }
-        warn "running '$clear'\n";
+        warn "running '$clear'\n" if $gDebug;
         system $clear;
     }
 }
